refactory : PostAPI AOP and Autoken Logging setting
