+++ src/main/java/Gdsc/web/oauth/token/AuthToken.java
@@ -92,7 +92,7 @@ public class AuthToken {
         } catch (UnsupportedJwtException e) {
             log.info("Unsupported JWT token.");
         } catch (IllegalArgumentException e) {
-            log.info("JWT token compact of handler are invalid.");
+            //log.info("JWT token compact of handler are invalid.");
         }
         return null;
     }
